requirements.txt
openai>=1.0.0
aiofiles>=23.0.0
telethon>=1.29.0
python-docx>=0.8.11
# Existing dependencies
google-generativeai>=0.3.0
tqdm>=4.64.0 